feat: add import endpoints for API v3
